// src/store/auth/action-types.ts
// Action enums
export enum ActionTypes {
  LOGIN = 'LOGIN',
  REGISTER = 'REGISTER',
  FORGOT_PASSWORD = 'FORGOT_PASSWORD',
  VERIFY_TOKEN = 'VERIFY_TOKEN',
  RESET_PASSWORD = 'RESET_PASSWORD',
  SWITCH_ACCOUNT = 'SWITCH_ACCOUNT',
  SEARCH_ACCOUNTS = 'SEARCH_ACCOUNTS',
  LOGOUT = 'LOGOUT',
  GET_CATEGORIES_ACCSETS = 'GET_CATEGORIES_ACCSETS',
}